refactor : Section과 Line에 대한 Request 분리
