+++ src/test/java/wooteco/subway/acceptance/template/LineRequest.java
@@ -5,7 +5,6 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
 import wooteco.subway.controller.dto.request.LineCreateRequestDto;
-import wooteco.subway.controller.dto.request.SectionRequestDto;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -36,19 +35,5 @@ public class LineRequest {
         params.put("downStationId", dto.getDownStationId().toString());
         params.put("distance", Integer.toString(dto.getDistance()));
         return params;
-    }
-
-    public static ExtractableResponse<Response> createSectionRequestAndReturnResponse(SectionRequestDto dto, Long lineId) {
-        Map<String, String> params = new HashMap<>();
-        params.put("downStationId", dto.getDownStationId().toString());
-        params.put("upStationId", dto.getUpStationId().toString());
-        params.put("distance", Integer.toString(dto.getDistance()));
-        return RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines/{lineId}/sections", lineId)
-                .then().log().all()
-                .extract();
     }
 }
